refactor(hal): use uintptr_t for hardware address enums, rename smc_generated to smc (#402)
* refactor(hal): use uintptr_t for hardware address enums, rename smc_generated to smc

Replace uint32_t with uintptr_t as the underlying type for all C23 typed enums
that hold RX72N hardware peripheral base addresses across 30 files. uintptr_t is
the semantically correct type for storing pointer-sized addresses, whereas uint32_t
expresses only integer width. On RX72N (32-bit) both are the same width, so this
is a correctness and documentation improvement with zero runtime cost.

Also renames src/boot/smc_generated/ and src/boot/inc/smc_generated/ to smc/ to
align with STAR coding standards (dropping the redundant _generated suffix).

Files changed:
- 28x rx_hal/inc/rx72n_*_regs.h: address enums uint32_t -> uintptr_t
- tests/mocks/rx72n_system_regs.h: same fix for mock address enums
- src/boot/smc/lowlvl.c: E1 debug port address enum uint32_t -> uintptr_t
- src/boot/smc_generated/ -> src/boot/smc/ (renamed, 5 source files)
- src/boot/inc/smc_generated/ -> src/boot/inc/smc/ (renamed, 7 header files)
- CMakeLists.txt: update include path from inc/smc_generated to inc/smc

* fix(hal): fix missed uintptr_t conversions in rx72n_flash_regs.h

Four address enums were missed in the initial pass because their hex
literals fall outside the 0x00080000-0x000FFFFF peripheral range:

- rx_faci_addresses_t (line 261): FACI flash registers at 0x007FE000
- rx_eepfclk_addresses_t (line 288): data flash freq register at 0x007FC040
- rx_uidr_addresses_t (line 833): Unique ID ROM registers at 0xFE7F7D90
- rx_flash_memory_addresses_t (line 892): code/data flash region addresses

All four enums contain values cast to volatile pointers in accessor
functions and must use uintptr_t, not uint32_t.

* docs(hal): add Doxygen block to rx_usb_addresses_t enum

* fix(hal): correct non-address enums and clean up global includes

- Revert mhiti_bits_t and mhitd_bits_t in rx72n_mpu_regs.h from uintptr_t
  to uint32_t; these hold register bitfields and shift amounts, not addresses
- Split rx_flash_memory_addresses_t in rx72n_flash_regs.h: keep uintptr_t for
  address constants, move sizes/blocks/program-units to new rx_flash_memory_layout_t
  (uint32_t)
- Remove redundant src/boot/inc/smc from global include_directories() in
  CMakeLists.txt; it is already covered by target_include_directories()

* docs(hal): fix uint32_t in Doxygen example for uintptr_t address arithmetic

The @code example in rx72n_port_regs.h showed uint32_t variables holding
the result of uintptr_t address arithmetic, which would silently truncate
on 64-bit hosts. Updated examples to use uintptr_t consistently.

* docs: require uintptr_t for hardware address enums in CLAUDE.md and CodeRabbit

Update the enum type selection guide in CLAUDE.md to explicitly require
uintptr_t for all hardware memory-mapped address enums, and add a
dedicated "Hardware Address Enums" enforcement section to both C firmware
path blocks in .coderabbit.yaml.

On the 32-bit RX72N uintptr_t == uint32_t so there is no runtime
difference, but on the 64-bit x86_64 unit-test host uint32_t silently
truncates pointer casts and produces wrong addresses. PR 402 and the
follow-up audit confirmed all existing address enums already use
uintptr_t; these changes codify that as a mandatory rule so CodeRabbit
will reject any new address enum that regresses to uint32_t.

* docs(hal): expand Doxygen on flash memory address and layout enums

rx_flash_memory_addresses_t and rx_flash_memory_layout_t were missing
@details, @note, @invariant, @code, and @see blocks required by the
project Doxygen policy. Add full documentation covering address-space
layout, split-block erase geometry, program-unit alignment constraints,
and cross-references to the FCU accessor functions.

+++ star-rx72n-firmware/libs/rx_hal/inc/rx72n_cmt_regs.h
@@ -182,7 +182,7 @@ extern "C" {
  *
  * @note CMSTR registers are at base, channel registers follow at +2 offset.
  */
-typedef enum : uint32_t {
+typedef enum : uintptr_t {
   k_cmt_ctrl_base_addr = 0x00088000, /**< CMT control register base (CMSTR0/1) */
   k_cmt0_base_addr     = 0x00088002, /**< CMT0 registers (CMCR, CMCNT, CMCOR) */
   k_cmt1_base_addr     = 0x00088008, /**< CMT1 registers (CMCR, CMCNT, CMCOR) */
